+++ factoryFont/newFontWithStyleVariantWeight.go
@@ -36,12 +36,12 @@ import (
 //     fontWeight.KLighter ou fontWeight.K100 a fontWeight.K900
 //     density: Por favor, veja o elemento densidade
 //     iDensity: Por favor, veja o elemento densidade
-func NewFontWithStyleVariantWeight(size int, sizeUnit string, fontColor color.RGBA, family fontFamily.FontFamily, style fontStyle.FontStyle, variant fontVariant.FontVariant, weight fontWeight.FontWeight, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) font.Font {
+func NewFontWithStyleVariantWeight(size float64, sizeUnit string, fontColor color.RGBA, family fontFamily.FontFamily, style fontStyle.FontStyle, variant fontVariant.FontVariant, weight fontWeight.FontWeight, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) font.Font {
 	densityCalc := iDensity
 	densityCalc.SetDensityFactor(density)
 
 	densityCalc.Set(size)
-	size = densityCalc.Int()
+	size = densityCalc.Float64()
 
 	f := font.Font{
 		Size:     size,
