+++ factoryText/nweTextOutlineOnlyWithFont.go
@@ -8,15 +8,15 @@ import (
 	"github.com/helmutkemper/iotmaker.platform/abstractType/text"
 )
 
-func NewTextOutlineOnlyWithFont(platform iotmaker_platform_IDraw.IDraw, shadow iotmaker_platform_IDraw.IFilterShadowInterface, gradient iotmaker_platform_IDraw.IFilterGradientInterface, color interface{}, labelFont font.Font, label string, x, y int, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) text.Text {
+func NewTextOutlineOnlyWithFont(platform iotmaker_platform_IDraw.IDraw, shadow iotmaker_platform_IDraw.IFilterShadowInterface, gradient iotmaker_platform_IDraw.IFilterGradientInterface, color interface{}, labelFont font.Font, label string, x, y float64, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) text.Text {
 	densityCalc := iDensity
 	densityCalc.SetDensityFactor(density)
 
 	densityCalc.Set(x)
-	x = densityCalc.Int()
+	x = densityCalc.Float64()
 
 	densityCalc.Set(y)
-	y = densityCalc.Int()
+	y = densityCalc.Float64()
 
 	ik := genericTypes.Ink{}
 	ik = genericTypes.NewInc(ik, 0, color, shadow, gradient, density, iDensity)
