+++ factoryText/nweTextOutlineOnlyWithMaxWidth.go
@@ -7,15 +7,15 @@ import (
 	"github.com/helmutkemper/iotmaker.platform/abstractType/text"
 )
 
-func NewTextOutlineOnlyWithMaxWidth(platform iotmaker_platform_IDraw.IDraw, shadow iotmaker_platform_IDraw.IFilterShadowInterface, gradient iotmaker_platform_IDraw.IFilterGradientInterface, color interface{}, label string, x, y, maxWidth int, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) text.Text {
+func NewTextOutlineOnlyWithMaxWidth(platform iotmaker_platform_IDraw.IDraw, shadow iotmaker_platform_IDraw.IFilterShadowInterface, gradient iotmaker_platform_IDraw.IFilterGradientInterface, color interface{}, label string, x, y, maxWidth float64, density interface{}, iDensity iotmaker_platform_coordinate.IDensity) text.Text {
 	densityCalc := iDensity
 	densityCalc.SetDensityFactor(density)
 
 	densityCalc.Set(x)
-	x = densityCalc.Int()
+	x = densityCalc.Float64()
 
 	densityCalc.Set(y)
-	y = densityCalc.Int()
+	y = densityCalc.Float64()
 
 	ik := genericTypes.Ink{}
 	ik = genericTypes.NewInc(ik, 0, color, shadow, gradient, density, iDensity)
