+++ mouse/typePointerCollisionFunction.go
@@ -1,3 +1,3 @@
 package mouse
 
-type PointerCollisionFunction func(x, y int) bool
+type PointerCollisionFunction func(x, y float64) bool
