+++ mouse/typePointerEventFunction.go
@@ -1,3 +1,3 @@
 package mouse
 
-type PointerEventFunction func(x, y int, collision bool)
+type PointerEventFunction func(x, y float64, collision bool)
